update make_rankers to do ranking and sorting on classes

+++ src/scripts/file_parsing.py
@@ -13,6 +13,45 @@ def is_int(s):
     except:
         return False
 
+### CategoryClass Class
+### Contains a list of categories which make up
+### a CategoryClass. E.g. The two categories
+### 'sports' and 'health and fitness' could make
+### up a single class
+
+class CategoryClass:
+
+### Constructor
+
+    def __init__(self, categories):
+        self.classes = categories
+
+### Static initalizers to parse in the classes from
+### a csv file
+
+    @staticmethod
+    def parse(file_path):
+        classes = []
+        with open(file_path, 'r') as file:
+            for l in file.readlines():
+                l = l.strip()
+                c = CategoryClass(l.split(','))
+                classes.append(c)
+        return classes
+
+    @staticmethod
+    def category_to_ranker(category):
+        return category.replace(' ', '').replace('/', '_')
+
+### Function to determine if a category is within
+### this class
+
+    def in_class(self, category):
+        return category in self.classes
+
+### End CategoryClass
+
+
 
 ### Article class
 ### Contains all the files of a single article
@@ -112,14 +151,18 @@ class Articles:
 
 ### Constructor
 
-    def __init__(self, path):
+    def __init__(self, path, par=True):
         pathes = [join(path, f) for f in listdir(path) if isdir(join(path, f))]
-
-        cpu_ct = mp.cpu_count()
-        pool = mp.Pool(cpu_ct)
-        rtn = pool.map(Article, pathes)
-        pool.terminate()
-        self.articles = [a for a in rtn if a.valid]
+        if par:
+            cpu_ct = mp.cpu_count()
+            pool = mp.Pool(cpu_ct)
+            self.articles = pool.map(Article, pathes)
+            pool.terminate()
+            
+        else:
+            self.articles = [Article(a) for a in pathes]
+            
+        self.articles = [a for a in self.articles if a.valid]
 
 ### End Constructor
 
@@ -135,13 +178,13 @@ class Articles:
                         'type' : r['type'],
                         'sentence' : r['sentence'],
                         'score' : float(r['score']),
-                        'category' : cats[-1] if len(cats) > 0 else ''
+                        'category' : cats[0] if len(cats) > 0 else ''
                 }
                 if rel['type'] not in types:
                     types[rel['type']] = []
                 types[rel['type']].append(rel)
 
-            return types
+        return types
 
     def category_counts(self):
         cats_dict = { }
@@ -160,12 +203,32 @@ class Articles:
             set_cats.add(k)
         return set_cats
 
-### Static functions
+    def first_level_categories(self):
+        set_cats = set()
+        for a in self.articles:
+            cats = a.categories()
+            if len(cats) > 0: 
+                set_cats.add(cats[0])
+        return set_cats
 
-    @staticmethod
-    def CategoryToRanker(category):
-        return category.replace(' ', '').replace('/', '_')
+    def filter_categories(self, classes):
+        types = self.relation_types()
+        by_class = { }
+        for k in types:
+            for rel in types[k]:
+                cat = rel['category']                
+                for c in classes:
+                    if c.in_class(cat):
+                        if c not in by_class:
+                            by_class[c] = []
+                        by_class[c].append(rel)
 
-### End static functions
+        return by_class
+
+    def filter_sort_categories(self, classes):
+        cat_class = self.filter_categories(classes)
+        for k in cat_class:
+            cat_class[k] = sorted(cat_class[k], key=lambda x: -x['score'])
+        return cat_class
 
 ### End Articles class
